% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinflip.R
\name{coinflip}
\alias{coinflip}
\title{coinflip}
\usage{
coinflip(flips, str.out = FALSE, head = 0)
}
\arguments{
\item{flips}{number of coins to flip}

\item{str.out}{should the result be a value of 0 or 1 or a string of Kopf or Zahl}

\item{head}{which value represents head, should be a value of \code{c(0, 1)}}
}
\value{
a vector of class gamble
}
\description{
Simulates flipping of a coin
}
\examples{
coinflip(10, str.out = TRUE)

}
