% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{throw_dice}
\alias{throw_dice}
\title{throw_dice}
\usage{
throw_dice(throws = 1, sides = 6)
}
\arguments{
\item{throws}{number of throws of dice}

\item{sides}{number of dice sides}
}
\value{
a vector of class gamble
}
\description{
Simulates rolling of dice
}
\examples{
throw_dice(throws = 1, sides = 6)

}
